requirements.txt
python-telegram-bot==20.7
google-api-python-client==2.108.0
yt-dlp==2023.12.30
python-dotenv==1.0.0
aiohttp==3.9.1
